It works
the calculations work, and the buttons work. Need to take out bugtesting feedback, and I plan to do some polishing touches for weird inputs and make the interface nicer maybe. Also seems it ~kinda~ works in landscape mode (doesnt scroll, but layout and buttons keep), so might try to look into that next.